Rename (and overwrite) robot.py to rover.py
